requirements.txt
PyGithub>=1.57.0
PyYAML>=6.0.1
tomli>=2.0.0;python_version<"3.11"
